fully npmify
- update tests to mocha
