+++ README.md
@@ -1,22 +1,24 @@
 # Knockout Init Component
 
-> This is a simple knockout component that will automatically create observables from arbitrary json.
+> This is a simple knockout component which will automatically create observables from arbitrary json.
 
-It uses [knockout.moment](https://github.com/ArchonInfoSys/knockout-moment), [knockout.money](https://github.com/ArchonInfoSys/knockout-money), and [knockout.integer](https://github.com/ArchonInfoSys/knockout-integer) to automatically extend the observables with the specified names.
+It uses [knockout.moment](https://github.com/civicsource/knockout.moment), [knockout.money](https://github.com/civicsource/knockout.money), and [knockout.integer](https://github.com/civicsource/knockout.integer) to automatically extend the observables with the specified names.
 
-##Install with [Bower](http://bower.io/)
+## Installation
 
 ```
-bower install knockout-init
+npm install knockout-init
 ```
 
 Then add `knockout.init.js` to your project.
 
-##How to Use
+## Usage
 
-Include the script on your page (either via a normal script tag or via an AMD loader). Then bind it to a button or a link.
+Include the script in your bundle, then:
 
 ```js
+var koInit = require("knockout.init");
+
 var dataFromServer = {
 	createdOn: "May 15, 2014",
 	expiresOn: "May 15, 2037",
@@ -37,18 +39,17 @@ console.log(myObj.numberOfThings.formatted()); //4,700
 console.log(myObj.price.formatted()); //$570,000.00
 ```
 
-##How to Run Tests
+## How to Run Tests
 
 Clone this repository and run:
 
 ```
 npm install
-bower install
-grunt
+npm run test
 ```
 
-You will need to have `grunt-cli` installed globally if you don't already have it. Install it if you don't:
+You will want to have `mocha` installed globally if you don't already have it. Install it if you don't:
 
 ```
-npm install grunt-cli -g
+npm install mocha -g
 ```
