--- bower.json
@@ -1,15 +1,0 @@
-{
-  "name": "knockout-init",
-  "version": "1.1.0",
-  "main": "knockout.init.js",
-  "ignore": [
-    "node_modules",
-    "bower_components"
-  ],
-  "dependencies": {
-    "knockout": "~3.1.0",
-    "knockout-moment": "~1.0.0",
-    "knockout-integer": "~1.0.0",
-    "knockout-money": "~1.0.0"
-  }
-}
