+++ package.json
@@ -1,18 +1,25 @@
 {
-  "name": "knockout-init",
-  "version": "1.1.0",
+  "name": "knockout.init",
+  "version": "1.0.0",
   "description": "This is a simple knockout component that will automatically create observables from arbitrary json",
-  "main": "app.js",
+  "main": "knockout.init.js",
   "scripts": {
-    "test": "grunt test"
+    "lint": "eslint . --format=unix",
+    "test": "mocha"
   },
-  "private": true,
-  "author": "Archon",
+  "author": "CivicSource",
   "license": "ISC",
+  "dependencies": {
+    "knockout.integer": "2.x",
+    "knockout.moment": "2.x",
+    "knockout.money": "2.x"
+  },
   "devDependencies": {
-    "grunt": "^0.4.5",
-    "grunt-contrib-jshint": "^0.10.0",
-    "grunt-contrib-jasmine": "^0.5.3",
-    "grunt-template-jasmine-requirejs": "^0.1.10"
+    "chai": "3.x",
+    "knockout": "3.x",
+    "mocha": "2.x"
+  },
+  "peerDependencies": {
+    "knockout": "3.x"
   }
 }
